Add files via upload
This is the code used to send a stop and go signal over WiFi from processing to tell the buggy to start following the track and to stop moving completely. An ultrasonic sensor was used to detect objects in-front of the buggy and stop at a certain distance. This stopping distance is then graphed in processing.