Silver Buggy
The same as the Bronze buggy challenge but with code built in so that the buggy will match the speed of an object moving in front. The UI also has an addition of a rotating buggy which shows the direction that the buggy is turning at a certain moment in time. Processing code was split into tabs.